Change how enums are hashed
Instead of using oid when hashing an enum value, instead look up the
enumsortorder and pass that into the hash function. This is necessary
because enum OIDs will be rewritten when restored to a new database,
and this will result in non-deterministic hashing.  The hash value is
used to distribute rows to various segments and so restoring data will
fail for any tables distributed by an enum column if hashed by OID.